add label-no-colon warning (default off)
This is a simplistic approach which will result in duplicate warnings
if both label-no-colon and label-orphan are enabled. A more advanced
approach would require functions such as nasm_is_suppressed_warning
and nasm_is_warning_error, to figure out which warning to emit and
whether the warning should be an error or not.